1-4 works... somewhat. can't do any more
The nodes Motion Match and StrideR don't seem to work in S&box.
For Motion Match you can't set the Parameters so it just won't find the right animation and sit in Idle.
for StrideR.. it just straight up crashes the game lmao.

Turn Helper also doesn't seem to 100% be working like it is suppose to. and we also can't set look target for Animgraph afaik. unless its a custom parameter.